move trade file into agent
