+++ lib/agent0/agent0/hyperdrive/agent/__init__.py
@@ -2,5 +2,14 @@
 
 from .hyperdrive_actions import HyperdriveActionType, HyperdriveMarketAction
 from .hyperdrive_agent import HyperdriveAgent
+from .hyperdrive_trades import (
+    add_liquidity_trade,
+    close_long_trade,
+    close_short_trade,
+    open_long_trade,
+    open_short_trade,
+    redeem_withdraw_shares_trade,
+    remove_liquidity_trade,
+)
 from .hyperdrive_wallet import HyperdriveWallet, HyperdriveWalletDeltas, Long, Short
 from .trade_result import TradeResult, TradeStatus
