+++ lib/agent0/agent0/hyperdrive/exec/__init__.py
@@ -8,15 +8,6 @@ from .execute_agent_trades import (
 )
 from .fund_agents import async_fund_agents
 from .get_agent_accounts import get_agent_accounts, set_max_approval
-from .hyperdrive_trades import (
-    add_liquidity_trade,
-    close_long_trade,
-    close_short_trade,
-    open_long_trade,
-    open_short_trade,
-    redeem_withdraw_shares_trade,
-    remove_liquidity_trade,
-)
 from .run_agents import build_wallet_positions_from_data, setup_and_run_agent_loop
 from .setup_experiment import setup_experiment
 from .trade_loop import get_wait_for_new_block, trade_if_new_block
