+++ lib/agent0/agent0/hyperdrive/policies/random.py
@@ -11,8 +11,7 @@ from fixedpointmath import FixedPoint
 
 from agent0.base import Trade
 from agent0.hyperdrive import HyperdriveActionType, HyperdriveMarketAction
-from agent0.hyperdrive.crash_report import build_crash_trade_result, log_hyperdrive_crash_report
-from agent0.hyperdrive.exec import (
+from agent0.hyperdrive.agent import (
     add_liquidity_trade,
     close_long_trade,
     close_short_trade,
@@ -21,6 +20,7 @@ from agent0.hyperdrive.exec import (
     redeem_withdraw_shares_trade,
     remove_liquidity_trade,
 )
+from agent0.hyperdrive.crash_report import build_crash_trade_result, log_hyperdrive_crash_report
 
 from .hyperdrive_policy import HyperdriveBasePolicy
